Create InstantViewOfBigBang.cpp
a very good problems for someone who is novice in graph algorithm. Introduces a lot of different and new techniques that one might not be able to implement at an early stage all at once, while at the same time teaching the importance of the different time complexity of different algorithms and using the appropriate one where necessary.